LayoutTests/http/tests/blink/sendbeacon/beacon-cross-origin.https.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verify navigator.sendBeacon() mixed content checking.");

window.jsTestIsAsync = true;

function test() {
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    shouldBeFalse('navigator.sendBeacon("http://example.test:8000/blink/sendbeacon/resources/save-beacon.py?name=cross-origin", "CrossOrigin");');
    finishJSTest();
}
</script>
</head>
<body onload="test();">
</body>
</html>
