LayoutTests/http/tests/blink/sendbeacon/beacon-cross-origin.https-expected.txt

Verify navigator.sendBeacon() mixed content checking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL navigator.sendBeacon("http://example.test:8000/blink/sendbeacon/resources/save-beacon.php?name=cross-origin", "CrossOrigin"); should be false. Was true.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
